+++ src/globals.d.ts
@@ -1,5 +1,8 @@
 declare const __DECK_SHELVES_ENABLE_HOME_PATCH__: boolean;
 declare const __DEV__: boolean;
+declare const __QA_FIRST_RUN__: boolean;
+declare const __QA_QAM_ERROR__: boolean;
+declare const __QA_SHELF_ERROR__: boolean;
 
 interface Window {
   SP_REACT?: any;